+++ src/__init__.py
@@ -2,9 +2,10 @@
 src
 Samuel Koller
 Created: 17 October 2024
-Updated: 17 October 2024
+Updated: 18 October 2024
 
 Contains all source code used for simulating a running elevator.
 """
 
 from .classes import *
+from .utils import *