+++ src/utils/constants.py
@@ -1,0 +1,12 @@
+"""
+constants.py
+Samuel Koller
+Created: 18 October 2024
+Updated: 19 October 2024
+
+Configurable constants to edit the default application.
+"""
+
+TOP_FLOOR = 20  # Default: 20
+PLAYBACK_SPEED = 5  # Default: 1
+TIMEOUT_TIME = 10  # Default: 10